Parse first kind of subpacket

+++ day16/src/bit_stream.rs
@@ -1,4 +1,4 @@
-const BITS: [[u32; 4]; 16] = [
+const BITS: [[usize; 4]; 16] = [
     [0, 0, 0, 0],
     [0, 0, 0, 1],
     [0, 0, 1, 0],
@@ -17,7 +17,7 @@ const BITS: [[u32; 4]; 16] = [
     [1, 1, 1, 1],
 ];
 
-pub fn bits(input: &str) -> Vec<u32> {
+pub fn bits(input: &str) -> Vec<usize> {
     (0..input.len())
         .map(|i| usize::from_str_radix(&input[i..i + 1], 16).unwrap())
         .flat_map(|hex| BITS[hex])
